Module switchboard_attestation_client::sgx_dcap_quoteverify_rs

Re-exports

pub use qvl_sys::quote3_error_t;

pub use qvl_sys::sgx_ql_request_policy_t;

pub use qvl_sys::sgx_ql_qv_supplemental_t;

pub use qvl_sys::sgx_ql_qve_collateral_t;

pub use qvl_sys::tdx_ql_qve_collateral_t;

pub use qvl_sys::sgx_ql_qv_result_t;

pub use qvl_sys::sgx_ql_qe_report_info_t;

pub use qvl_sys::sgx_qv_path_type_t;

pub use qvl_sys::tee_supp_data_descriptor_t;

Functions

sgx_qv_get_quote_supplemental_data_size

Get SGX supplemental data required size.

sgx_qv_set_enclave_load_policy

When the Quoting Verification Library is linked to a process, it needs to know the proper enclave loading policy. The library may be linked with a long lived process, such as a service, where it can load the enclaves and leave them loaded (persistent). This better ensures that the enclaves will be available upon quote requests and not subject to EPC limitations if loaded on demand. However, if the Quoting library is linked with an application process, there may be many applications with the Quoting library and a better utilization of EPC is to load and unloaded the quoting enclaves on demand (ephemeral). The library will be shipped with a default policy of loading enclaves and leaving them loaded until the library is unloaded (PERSISTENT). If the policy is set to EPHEMERAL, then the QE and PCE will be loaded and unloaded on-demand. If either enclave is already loaded when the policy is change to EPHEMERAL, the enclaves will be unloaded before returning.

sgx_qv_set_path

Set the full path of QVE and QPL library. The function takes the enum and the corresponding full path.

sgx_qv_verify_quote

Perform SGX ECDSA quote verification.

tdx_qv_get_quote_supplemental_data_size

Get TDX supplemental data required size.

tdx_qv_verify_quote

Perform TDX ECDSA quote verification.

tee_get_supplemental_data_version_and_size

Get supplemental data latest version and required size, support both SGX and TDX

tee_qv_get_collateral

Get quote verification collateral.

tee_verify_quote

Perform quote verification for SGX and TDX This API works the same as the old one, but takes a new parameter to describe the supplemental data (p_supp_data_descriptor)